Started migration to WebAssembly!!! 🥳
All Vector3D and Matrix4x4 function currently being used have been migrated to AssemblyScript and are now compiled to WebAssembly before being used!

I also added tests for all migrated functions to ensure I didn't fudge anything up 😅

There's more to do (the Quaternion function still need to be migrated, and the view projection matrix calculation should be migrated) - but this is an awesome start!

I'm so excited; I've wanted to work with WebAssembly for a *long* time - and it's finally happening! 🤘🏻