I got the sphere rendering! 🥳🤘🏻💯
Well ... not *completely* 😅

The coordinates for each vertex are just the corners of the voxel used to test for the sphere surface - so they're not in the correct position.

This means that what's rendered is a 3D octagon, not an actual sphere ... but 98% of all the logic required to render the actual sphere is in place!

This is also the most brute-force approach possible (if you look at the diff you'll see that it takes ~1k lines of code to add all the necessary vertices).

I'll optimize/refactor things in upcoming commits.

I'm really proud of what I've accomplished with this 😊